Updated AmanithSVG and added support for notched devices. Refactored code

+++ core/src/com/azokh/blackbox/ui/element/button/BBListenerButton.java
@@ -8,12 +8,12 @@ public class BBListenerButton extends BBButtonText
 {
     private final BBListener ffListener;
 
-    public BBListenerButton(int id, int x, int y, String text, BitmapFont font, BBListener ffListener) {
+    public BBListenerButton(int id, float x, float y, String text, BitmapFont font, BBListener ffListener) {
         super(id, x, y, text, font, Align.left);
         this.ffListener = ffListener;
     }
 
-    public BBListenerButton(int id, int x, int y, String text, BitmapFont font, BBListener ffListener, int align) {
+    public BBListenerButton(int id, float x, float y, String text, BitmapFont font, BBListener ffListener, int align) {
         super(id, x, y, text, font, align);
         this.ffListener = ffListener;
     }
