+++ core/src/com/azokh/blackbox/ui/element/button/BBListenerButtonTexture.java
@@ -6,12 +6,12 @@ import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
 public class BBListenerButtonTexture extends BBButtonTexture{
     private final BBListener ffListener;
 
-    public BBListenerButtonTexture(int id, int x, int y, Texture texture1, Texture texture2, BBListener ffListener) {
+    public BBListenerButtonTexture(int id, float x, float y, Texture texture1, Texture texture2, BBListener ffListener) {
         super(id, x, y, texture1, texture2);
         this.ffListener = ffListener;
     }
 
-    public BBListenerButtonTexture(int id, int x, int y, Texture texture, BBListener ffListener) {
+    public BBListenerButtonTexture(int id, float x, float y, Texture texture, BBListener ffListener) {
         super(id, x, y, texture);
         this.ffListener = ffListener;
     }
